+++ CHANGES.rst
@@ -15,6 +15,8 @@
 
 - Implemented support, tests, and maker utility for Super Bias reference files. [#45]  
 
+- Created maker utility and tests for wfi photom reference files. [#43]
+
 - Added support, tests, and maker utility for Pixel Area reference files. [#44]